backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0

boto3>=1.35

stripe==13.1.0

google-generativeai>=0.7.0
openpyxl>=3.1
reportlab>=4.2
django-filter>=24.2
firebase-admin>=6.5.0
gunicorn>=22.0
